Update image pointer, try to fix AOS
This changes the QCA logo in the about page from the SVG which dropped
the QC (bad svg, not respecting "foreign artists" from Illustrator).

This also attempts to fix the images on the home page not coming in
correctly due to what appears to be the Animate On Scroll (AOS) plugin
not playing nicely with the `body{overflow-x: hidden;}` flag.
See https://github.com/michalsnik/aos/issues/110 for more details.

Fixes #15