Fix Carousel image size failings
Owl Carousel (the plugin controlling the cycling images) failing to get correct image size due to the dynamic page creation of Hugo. This was most problematic on the About and Get Started pages.

The Carousel `div` calculated image size was calculated before content was
populated, so fell to default of `width: 100%`, or the full screen, but was still contained to the column. The images were then populated and stretched, but contained in the box.

Triggering a resize event caused a re-calculation with the now populated images, which fixed the
problem.

This PR triggers a single Carousel refresh once the window is loaded, after content population,
to get the sizes right.

See also:
https://github.com/OwlCarousel2/OwlCarousel2/issues/461#issuecomment-330080323
https://github.com/OwlCarousel2/OwlCarousel2/issues/2054#issuecomment-328081874

Related to but not fixed by Owl Carousel 2.3.4, which isn't their fault just oddities of how Hugo
constructs its pages I think.