Make URL links look normal (blue)
Small change to the `_topography.scss` file fixes URL links to
actually look like URL links (blue) by removing the formatting applied
to them alongside the other types.

I think there is more we would have to add to ensure they got things
like underline and turned purple after clicking, but this is a small
thing we can change.

Not sure how this impacts other parts of the CSS or the banner at the
top, it might not, but I'm not familiar enough with CSS to really check
this.